-Added a Matrix class to be used in the transform class -Added a transform class that will serve as a root component to the GameObject class -modified the implementation of previous classes to avoid the use of namespaces as I was having some issues with my includes that were not worth the advantages namespaces offer.
next is modifying the sprite class or adding a new one, creating the scene class and testing the current added functionality to make sure no logic errors are present in the Math implementation.
After that I will focus on building the managers so the scene stack can be made.
Then come the colliders and collision handling.